requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.15.0
urllib3==2.5.0
